Grading
Good work,

I took off a few points in UI and functionality since your boxplot is the same code as your pointplot... Meaning it is not a boxplot.

Also in the future, for max/min you can actually write your own SQL statement that pulls out the min and max instead of a list of distinct values and then getting the min and max from there.